glsl-1.20: Add tests exposing some crashes with const arrays
I stumbled across these while implementing tessellation. In particular,
the first test is a cutdown version of the tessellation nop.shader_test.

The crash in array-initializers-1 is caused by the lowering of const arrays to
uniforms. The hidden uniforms for both are named 'constarray', and a
later phase in the linker gets horribly confused when it tries to look
them up by name.

I have not yet usefully characterized the crash in double-indirect-1,
other than that it is in the i965 backend rather than in the GLSL
compiler.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>